♻️ Migrate the Slack client codes into the new subpackage *client* and rename the modules to be more clear.
