+++ slack_mcp/client/factory.py
@@ -177,7 +177,7 @@ class DefaultSlackClientFactory(SlackClientFactory):
         AsyncWebClient
             Initialized Slack AsyncWebClient instance
         """
-        from slack_mcp.client_manager import get_client_manager
+        from slack_mcp.client.manager import get_client_manager
 
         manager = get_client_manager()
         return self.create_async_client(manager._default_token)
